优化通知入库消费者，添加sql预处理

+++ app/common/common.go
@@ -74,7 +74,7 @@ func MD5(text string) string {
 	return hex.EncodeToString(md.Sum(nil))
 }
 
-func GetHaseValue(uid int)int {
+func GetHaseValue(uid uint64)uint64 {
 	return (uid % 16)+1
 }
